### 2024-08-22 #7135 @kthyng
In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any).

I see that this has already been reviewed in `pyOpenSci` here: https://github.com/pyOpenSci/software-submission/issues/157
