### 2024-08-23 #7135 @jbousquin
Hi @kthyng and thank you! From editorialbot that invalid doi was maybe missing a dash (DOI: 10.1111/j.1752‐1688.2011.00573.x [link](https://onlinelibrary.wiley.com/doi/full/10.1111/j.1752-1688.2011.00573.x)), and tidy data should be https://doi.org/10.18637/jss.v059.i10. I'll work on a PR for those.
