### 2024-08-26 #7135 @jbousquin
Made some minor updates to paper/paper.bib on the branch paper_JOSS. The doi editorialbot identified as invalid did contain the '-' but it wasn't being encoded correctly. Confirmed the doi link works as it should now in the pdf generated by the workflow.
