### 2024-08-28 #7135 @kthyng
@jbousquin a few paper comments (still need to read through in detail)
* the parenthetical references aren't in there correctly. You'll want to create a list of them rather than listed one, then another, each with parentheses. Here is an example paper: https://joss.readthedocs.io/en/latest/example_paper.html#example-paper
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "r" is not capitalized, but please check others.
* You and I are the only people that will be reading your paper so please check everything carefully! Usually reviewers and an editor would also be checking, but your repo was already reviewed so the paper will only be checked by us.
