### 2024-09-10 #7135 @kthyng
@jbousquin Can you check through your paper? I can also wait and start the review issue before we go into these details if you prefer.
