### 2024-09-30 #7305 @editorialbot
[REVIEW]: harmonize-wq: Standardize, clean and wrangle Water Quality Portal data into more analytic-ready formats
**Submitting author:** @jbousquin (<a href="http://orcid.org/0000-0001-5797-4322">Justin Bousquin</a>)
**Repository:** https://github.com/USEPA/harmonize-wq
**Branch with paper.md** (empty if default branch): paper_JOSS
**Version:** 0.5.0
