### 2024-10-02 #7305 @kthyng
@jbousquin
* is v0.5.0 still correct for this publication?
* Can you update the title and author list in Zenodo to match your JOSS publication?
