### 2024-10-03 #7305 @kthyng
How the paper file is arranged in the archive or if it is the most up to date isn't important since the words will be present in the paper itself, so no need from our perspective to do a new release for any changes in the paper.
