### 2024-10-09 #7305 @jbousquin
I think I was able to edit the authors (moved other types of contributors from creators to contributors) and title, everything else about that zenodo archive is auto-populated from github. Please let me know if any other fields need to be updated or anything doesn't match as it should.
